requirements.txt
AwesomeTkinter==2021.11.8
Babel==2.15.0
CTkMessagebox==2.5
customtkinter==5.2.2
darkdetect==0.8.0
distlib==0.3.8
filelock==3.15.1
greenlet==3.0.3
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
python-bidi==0.4.2
schedule==1.2.2
setuptools==70.0.0
six==1.16.0
tkcalendar==1.6.1
ttkbootstrap==1.10.1
typing_extensions==4.12.2
virtualenv==20.26.2
wheel==0.43.0
